+++ packages/xchain-polkadot/jest.setup.js
@@ -1,5 +1,0 @@
-const mock = require('mock-socket')
-WebSocket = mock.WebSocket
-
-jest.setTimeout(30000)
-console.log = jest.fn()