+++ packages/xchain-thorchain-amm/src/utils/midgard.ts
@@ -117,7 +117,6 @@ export class Midgard {
    */
   async getScheduledOutboundValue(): Promise<CryptoAmount> {
     const path = 'v2/thorchain/queue'
-
     for (const baseUrl of this.config.midgardBaseUrls) {
       try {
         const { data } = await axios.get(`${baseUrl}${path}`)
@@ -162,6 +161,22 @@ export class Midgard {
       try {
         const data = (await api.getHealth()).data
         return +data.scannerHeight
+      } catch (e) {
+        console.error(e)
+      }
+    }
+    throw Error(`Midgard not responding`)
+  }
+  /**
+   * Gets actions related to a txID
+   * @param txHash transaction id
+   * @returns Type Action array of objects
+   */
+  public async getActions(txHash: string): Promise<Action[]> {
+    for (const api of this.midgardApis) {
+      try {
+        const actions = (await api.getActions('', txHash)).data.actions
+        return actions
       } catch (e) {
         console.error(e)
       }