+++ packages/xchain-thorchain-amm/src/utils/thornode.ts
@@ -21,12 +21,15 @@ import {
   LTCChain,
   THORChain,
   TerraChain,
+  assetFromString,
 } from '@xchainjs/xchain-util'
 import axios from 'axios'
 import axiosRetry from 'axios-retry'
 
 import { DefaultChainAttributes } from '../chain-defaults'
 import { ChainAttributes } from '../types'
+
+import { Midgard } from './midgard'
 
 export type ThornodeConfig = {
   apiRetries: number
@@ -72,9 +75,11 @@ export class Thornode {
   private transactionsApi: TransactionsApi[]
   private queueApi: QueueApi[]
   private networkApi: NetworkApi[]
+  private midgard: Midgard
 
   constructor(network: Network = Network.Mainnet, config?: ThornodeConfig, chainAttributes = DefaultChainAttributes) {
     this.network = network
+    this.midgard = new Midgard(Network.Mainnet)
     this.config = config ?? defaultThornodeConfig[this.network]
     axiosRetry(axios, { retries: this.config.apiRetries, retryDelay: axiosRetry.exponentialDelay })
     this.transactionsApi = this.config.thornodeBaseUrls.map(
@@ -154,8 +159,6 @@ export class Thornode {
         return DOGEChain
       case 'TERRA':
         return TerraChain
-      case 'POLKA':
-        throw Error('Polkadot is not supported yet')
       default:
         throw Error('Unknown chain')
     }
@@ -172,53 +175,90 @@ export class Thornode {
   public async checkTx(inboundTxHash: string, sourceChain?: Chain): Promise<TxStatus> {
     let txStatus: TxStatus = { stage: TxStage.INBOUND_CHAIN_UNCONFIRMED, seconds: 0 }
     const txData = await this.getTxData(inboundTxHash)
-    const scheduledQueue = await this.getscheduledQueue()
-    const scheduledQueueItem = scheduledQueue?.find((item: TxOutItem) => item.in_hash === inboundTxHash)
-    const lastBlock = await this.getLastBlock()
-    const lastBlockHeight = lastBlock.find((obj) => {
-      return obj
-    })
-    console.log(txData.observed_tx?.tx)
-    if (txData.observed_tx?.tx == undefined) {
-      console.log(`Stage 1`)
-      txStatus = await this.checkObservedTx(txStatus, txData.observed_tx, sourceChain)
-      return txStatus
-    } else if (txData.observed_tx.status == ObservedTxStatusEnum.Incomplete) {
-      console.log(`Stage 2`)
-      txStatus = await this.checkObservedOnly(txStatus, txData.observed_tx, sourceChain)
-      return txStatus
-    } else if (txData.observed_tx.status == ObservedTxStatusEnum.Done) {
-      console.log(`Stage 3`)
-      txStatus = await this.checkOutboundQueue(
-        txStatus,
-        txData.observed_tx,
-        scheduledQueue,
-        scheduledQueueItem,
-        lastBlockHeight,
-      )
-      return txStatus
-    } else if (scheduledQueueItem?.height == lastBlockHeight?.thorchain) {
-      console.log(`Stage 4`)
-      txStatus = await this.checkOutboundTx(txStatus, scheduledQueueItem, lastBlockHeight)
-      return txStatus
-    } else {
-      console.log('here')
-      txStatus.stage = TxStage.OUTBOUND_CHAIN_UNCONFIRMED
-      txStatus.seconds = 0
+    const scheduledQueueItem = (await this.getscheduledQueue()).find(
+      (item: TxOutItem) => item.in_hash === inboundTxHash,
+    )
+
+    // Check to see if the transaction has been observed
+    if (txData.observed_tx == undefined) {
+      txStatus = await this.checkTxDefined(txStatus, sourceChain)
       return txStatus
     }
+    // If its scheduled and observed
+    if (scheduledQueueItem && txData.observed_tx) {
+      txStatus = await this.checkObservedOnly(txStatus, scheduledQueueItem, txData.observed_tx, sourceChain)
+    }
+    // Retrieve asset and chain from memo
+    const pool = txData.observed_tx.tx.memo?.split(`:`)
+    if (!pool) throw Error(`No pool found from memo`)
+    const getAsset = assetFromString(pool[1].toUpperCase())
+    const lastBlock = await this.getLastBlock()
+    const lastBlockHeight = lastBlock.find((obj) => obj.chain === getAsset?.chain)
+
+    // Check to see if its in the outbound queue
+    if (scheduledQueueItem) {
+      txStatus = await this.checkOutboundQueue(txStatus, scheduledQueueItem, lastBlockHeight)
+      // Check to see if there is an outbound wait
+      if (scheduledQueueItem?.height != undefined && txStatus.stage < TxStage.OUTBOUND_CHAIN_CONFIRMED) {
+        txStatus = await this.checkOutboundTx(txStatus, scheduledQueueItem, lastBlockHeight)
+      }
+      return txStatus
+    }
+
+    // If not in queue, outbound Tx sent // check synth // check it status == done
+    if (!scheduledQueueItem && getAsset) {
+      txStatus.stage = TxStage.OUTBOUND_CHAIN_UNCONFIRMED
+      if (getAsset?.synth) {
+        if (txData.observed_tx?.status == ObservedTxStatusEnum.Done) {
+          txStatus.stage = TxStage.OUTBOUND_CHAIN_CONFIRMED
+          txStatus.seconds = 0
+        } else {
+          txStatus.seconds = 6
+        }
+        return txStatus
+      }
+      if (txData.observed_tx?.status == ObservedTxStatusEnum.Done && getAsset.chain != Chain.THORChain) {
+        if (lastBlockHeight?.last_observed_in && txData.observed_tx.block_height) {
+          const chainblockTime = this.chainAttributes[getAsset.chain].avgBlockTimeInSecs
+          // get the chain attributes from Outbound Chain and work out time elapsed
+          const blockDifference = lastBlockHeight.last_observed_in - txData.observed_tx.block_height
+          const timeElapsed = blockDifference / chainblockTime
+          txStatus.seconds = timeElapsed > chainblockTime ? 0 : chainblockTime - timeElapsed
+        } else if (txData.observed_tx.tx.id && lastBlockHeight?.thorchain) {
+          const recordedAction = await this.midgard.getActions(txData.observed_tx.tx.id)
+          const recordedBlockheight = recordedAction.find((block) => {
+            return block
+          })
+          if (!recordedBlockheight) throw Error(`No height recorded`)
+          const chainblockTime = this.chainAttributes[getAsset.chain].avgBlockTimeInSecs
+          const blockDifference = lastBlockHeight?.thorchain - +recordedBlockheight.height
+          const timeElapsed =
+            (blockDifference * chainblockTime) / this.chainAttributes[getAsset.chain].avgBlockTimeInSecs
+          txStatus.seconds = timeElapsed > chainblockTime ? 0 : chainblockTime - timeElapsed
+          txStatus.stage = TxStage.OUTBOUND_CHAIN_CONFIRMED
+        }
+        return txStatus
+      } else {
+        txStatus.seconds = 0
+        txStatus.stage = TxStage.OUTBOUND_CHAIN_CONFIRMED
+      }
+      return txStatus
+    } else {
+      // case example "memo": "OUT:08BC062B248F6F27D0FECEF1650843585A1496BFFEAF7CB17A1CBC30D8D58F9C" where no asset is found its a thorchain tx. Confirms in ~6 seconds
+      txStatus.seconds = 0
+      txStatus.stage = TxStage.OUTBOUND_CHAIN_CONFIRMED
+    }
+    return txStatus
   }
 
   /** Stage 1  */
-  private async checkObservedTx(txStatus: TxStatus, txData?: ObservedTx, sourceChain?: Chain): Promise<TxStatus> {
+  private async checkTxDefined(txStatus: TxStatus, sourceChain?: Chain): Promise<TxStatus> {
     // If there is an error Thornode does not know about it. wait 60 seconds
     // If a long block time like BTC, can check or poll to see if the status changes.
-    if (!txData?.tx) {
-      if (sourceChain) {
-        txStatus.seconds = this.chainAttributes[sourceChain].avgBlockTimeInSecs
-      } else {
-        txStatus.seconds = 60
-      }
+    if (sourceChain) {
+      txStatus.seconds = this.chainAttributes[sourceChain].avgBlockTimeInSecs
+    } else {
+      txStatus.seconds = 60
     }
     return txStatus
   }
@@ -227,23 +267,21 @@ export class Thornode {
   // e.g. https://thornode.ninerealms.com/thorchain/tx/365AC447BE6CE4A55D14143975EE3823A93A0D8DE2B70AECDD63B6A905C3D72B
   private async checkObservedOnly(
     txStatus: TxStatus,
+    scheduledQueueItem: TxOutItem,
     observed_tx?: ObservedTx,
     sourceChain?: Chain,
   ): Promise<TxStatus> {
-    if (observed_tx?.tx.chain != undefined) {
+    if (observed_tx?.tx?.chain != undefined) {
       sourceChain = this.getChain(observed_tx.tx.chain)
     } else {
-      throw new Error(`Cannot get source chain ${observed_tx?.tx.chain}`)
+      throw new Error(`Cannot get source chain ${observed_tx?.tx?.chain}`)
     }
-    const scheduledOutbound = await this.getscheduledQueue()
-    const observedTxBlockHeight = scheduledOutbound.find((obj) => {
-      return obj.height
-    })
+
     //If observed by not final, need to wait till the finalised block before moving to the next stage, blocks in source chain
-    if (observed_tx?.block_height && sourceChain && observed_tx?.finalise_height && observedTxBlockHeight?.height) {
+    if (observed_tx?.block_height && observed_tx?.finalise_height && scheduledQueueItem.height) {
       if (observed_tx.block_height < observed_tx.finalise_height) {
         txStatus.stage = TxStage.CONF_COUNTING
-        const blocksToWait = observed_tx.finalise_height - observedTxBlockHeight?.height
+        const blocksToWait = observed_tx.finalise_height - scheduledQueueItem.height
         txStatus.seconds = blocksToWait * this.chainAttributes[sourceChain].avgBlockTimeInSecs
       } else if (observed_tx.status != ObservedTxStatusEnum.Done) {
         // processed but not yet full final, e.g. not 2/3 nodes signed
@@ -253,26 +291,20 @@ export class Thornode {
     }
     return txStatus
   }
-  /** Stage 3 */
+  /**
+   * Stage 3
+   * @param txStatus
+   * @param txData
+   * @param scheduledQueue
+   * @param scheduledQueueItem
+   * @param lastBlockHeight
+   * @returns
+   */
   private async checkOutboundQueue(
     txStatus: TxStatus,
-    txData: ObservedTx,
-    scheduledQueue: TxOutItem[],
     scheduledQueueItem?: TxOutItem,
     lastBlockHeight?: LastBlock,
   ): Promise<TxStatus> {
-    txStatus.stage = TxStage.OUTBOUND_QUEUED
-    if (scheduledQueue.length == 0 && txData.tx.memo) {
-      // it is not queued, outbound Tx sent
-      txStatus.stage = TxStage.OUTBOUND_CHAIN_UNCONFIRMED
-      const pool = txData.tx.memo.split(`:`, 2)
-      const ticker = pool[1].split(`.`, 1).toString()
-      const chain = this.getChain(ticker)
-      txStatus.seconds = this.chainAttributes[chain].avgBlockTimeInSecs
-      return txStatus
-    }
-    console.log(scheduledQueueItem?.height)
-    console.log(lastBlockHeight?.thorchain)
     // If the scheduled block is greater than the current block, need to wait that amount of blocks till outbound is sent
     if (scheduledQueueItem?.height && lastBlockHeight?.thorchain) {
       if (lastBlockHeight.thorchain < scheduledQueueItem?.height) {
@@ -282,6 +314,7 @@ export class Thornode {
         return txStatus
       } else {
         txStatus.stage = TxStage.OUTBOUND_CHAIN_UNCONFIRMED
+        return txStatus
       }
     }
     return txStatus
@@ -293,14 +326,13 @@ export class Thornode {
     lastBlockHeight?: LastBlock,
   ): Promise<TxStatus> {
     if (scheduledQueueItem?.height && lastBlockHeight?.thorchain) {
-      console.log(`here`)
       const blockDifference = scheduledQueueItem.height - lastBlockHeight?.thorchain
       const timeElapsed = blockDifference * this.chainAttributes[THORChain].avgBlockTimeInSecs
       if (blockDifference == 0) {
         // If Tx has just been sent, Stage 3 should pick this up really
         txStatus.stage = TxStage.OUTBOUND_CHAIN_UNCONFIRMED
         txStatus.seconds = this.chainAttributes[THORChain].avgBlockTimeInSecs
-      } else if (timeElapsed < this.chainAttributes[THORChain].avgBlockTimeInSecs) {
+      } else if (timeElapsed < txStatus.seconds) {
         // if the time passed since the outbound TX was sent is less than the outbound block time, outbound Tx unconfirmed, wait a bit longer.
         txStatus.stage = TxStage.OUTBOUND_CHAIN_UNCONFIRMED
         txStatus.seconds = this.chainAttributes[THORChain].avgBlockTimeInSecs - timeElapsed // workout how long to wait