+++ packages/xchain-thorchain-amm/CHANGELOG.md
@@ -1,3 +1,19 @@
+# v0.0.1.0-alpha3 (2022-08-08)
+
+## Remove
+
+- Remove `Polkadot` from chain defaults and chain references/switch cases.
+
+# v0.0.1.0-alpha2 (2022-07-25)
+
+## ADD
+
+- Add `Doswap()` function to thorchain-amm
+
+## Update
+
+- Changed values in `calcSwapNetworkFee` to suite the latest network fees
+
 # v0.0.1.0-alpha (2022-07-20)
 
 ## Module Created